rename react-redux hooks configuring/usage files
