+++ playground/src/connected/index.ts
@@ -1,4 +1,3 @@
 export * from './fc-counter-connected-bind-action-creators';
 export * from './fc-counter-connected-own-props';
 export * from './fc-counter-connected';
-export * from './fc-counter-connected-hooks';
