+++ playground/src/hooks/react-redux-hooks.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { FCCounter } from '../components';
 import { increment } from '../features/counters/actions';
-import { useSelector, useDispatch } from './fc-counter-connected-hooks';
+import { useSelector, useDispatch } from '../store/hooks';
 
 const FCCounterConnectedHooksUsage: React.FC = () => {
   const counter = useSelector(state => state.counters.reduxCounter);
